Fix security

+++ src/main/java/com/group3/kindergartenmanagementsystem/service/impl/CommentForChildServiceImpl.java
@@ -1,5 +1,6 @@
 package com.group3.kindergartenmanagementsystem.service.impl;
 
+import com.group3.kindergartenmanagementsystem.exception.ForbiddenAccessException;
 import com.group3.kindergartenmanagementsystem.exception.ResourceNotFoundException;
 import com.group3.kindergartenmanagementsystem.model.Child;
 import com.group3.kindergartenmanagementsystem.model.CommentForChild;
@@ -9,33 +10,40 @@ import com.group3.kindergartenmanagementsystem.repository.ChildRepository;
 import com.group3.kindergartenmanagementsystem.repository.CommentForChildRepository;
 import com.group3.kindergartenmanagementsystem.service.CommentForChildService;
 import com.group3.kindergartenmanagementsystem.service.CommentForTeacherService;
+import com.group3.kindergartenmanagementsystem.service.SecurityService;
+import lombok.AllArgsConstructor;
 import org.modelmapper.ModelMapper;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
 @Service
+@AllArgsConstructor
 public class CommentForChildServiceImpl implements CommentForChildService {
     CommentForChildRepository commentForChildRepository;
     ChildRepository childRepository;
+    SecurityService securityService;
     ModelMapper mapper;
-
-    public CommentForChildServiceImpl(CommentForChildRepository commentForChildRepository, ChildRepository childRepository, ModelMapper mapper) {
-        this.commentForChildRepository = commentForChildRepository;
-        this.childRepository = childRepository;
-        this.mapper = mapper;
-    }
 
     @Override
     public List<CommentForChildDTO> getAllCommentByChildId(Integer childId) {
-        List<CommentForChild> comments = commentForChildRepository.findAllByChildId(childId);
+        Child child = childRepository.findById(childId)
+                .orElseThrow(()-> new ResourceNotFoundException("Child", "id", childId));
+        if (!securityService.isParentOrTeacherOfChild(child))
+            throw new ForbiddenAccessException(HttpStatus.BAD_REQUEST,
+                    "User with username: "+securityService.getUsername()+" can't access to this child");
+        List<CommentForChild> comments = commentForChildRepository.findAllByChild(child);
         return comments.stream().map(this::mapToDTO).collect(Collectors.toList());
     }
 
     @Override
     public CommentForChildDTO getCommentById(Integer id) {
         CommentForChild comment = commentForChildRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Comment for child", "id", id));
+        if (!securityService.isParentOrTeacherOfChild(comment.getChild()))
+            throw new ForbiddenAccessException(HttpStatus.BAD_REQUEST,
+                    "User with username: "+securityService.getUsername()+" can't access to this child");
         return mapToDTO(comment);
     }
 
@@ -43,6 +51,9 @@ public class CommentForChildServiceImpl implements CommentForChildService {
     public CommentForChildDTO createNewComment(CommentForChildDTO commentForChildDTO) {
         Child child = childRepository.findById(commentForChildDTO.getChildId()).orElseThrow(
                 () -> new ResourceNotFoundException("Child", "id", commentForChildDTO.getChildId()));
+        if (!securityService.isParentOrTeacherOfChild(child))
+            throw new ForbiddenAccessException(HttpStatus.BAD_REQUEST,
+                    "User with username: "+securityService.getUsername()+" can't access to this child");
         CommentForChild comment = mapToEntity(commentForChildDTO);
         comment.setChild(child);
         return null;
@@ -52,6 +63,9 @@ public class CommentForChildServiceImpl implements CommentForChildService {
     public CommentForChildDTO updateCommentById(Integer id, CommentForChildDTO commentForChildDTO) {
         CommentForChild comment = commentForChildRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Comment for child", "id", id));
         Child child = childRepository.findById(commentForChildDTO.getChildId()).orElseThrow(() -> new ResourceNotFoundException("Child", "id", id));
+        if (!securityService.isParentOrTeacherOfChild(child))
+            throw new ForbiddenAccessException(HttpStatus.BAD_REQUEST,
+                    "User with username: "+securityService.getUsername()+" can't access to this child");
         comment.setComment(comment.getComment());
         comment.setPostMonth(commentForChildDTO.getPostMonth());
         comment.setChild(child);
@@ -62,6 +76,9 @@ public class CommentForChildServiceImpl implements CommentForChildService {
     @Override
     public String deleteCommentById(Integer id) {
         CommentForChild comment = commentForChildRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Comment for child", "id", id));
+        if (!securityService.isParentOrTeacherOfChild(comment.getChild()))
+            throw new ForbiddenAccessException(HttpStatus.BAD_REQUEST,
+                    "User with username: "+securityService.getUsername()+" can't access to this child");
         commentForChildRepository.delete(comment);
         return "Comment deleted successfully";
     }
